PB/gym_be/requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
defusedxml==0.6.0
idna==2.10
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.2
python-dateutil==2.8.2
Pillow==9.3
django-cors-headers==3.13.0

